+++ sdk/attestation/attestation/test/public/attestationTests.spec.ts
@@ -224,10 +224,10 @@ describe("[AAD] Attestation Client", function() {
         }
       );
 
-      assert.isNotNull(attestationResult.value.sgxCollateral);
-      assert.isUndefined(attestationResult.value.runtimeClaims);
-      expect(attestationResult.value.enclaveHeldData?.length).is.equal(binaryRuntimeData.length);
-      expect(attestationResult.value.enclaveHeldData).to.deep.equal(binaryRuntimeData);
+      assert.isNotNull(attestationResult.body.sgxCollateral);
+      assert.isUndefined(attestationResult.body.runtimeClaims);
+      expect(attestationResult.body.enclaveHeldData?.length).is.equal(binaryRuntimeData.length);
+      expect(attestationResult.body.enclaveHeldData).to.deep.equal(binaryRuntimeData);
 
       assert(attestationResult.token, "Expected a token from the service but did not receive one");
     }
@@ -240,15 +240,15 @@ describe("[AAD] Attestation Client", function() {
         }
       );
 
-      assert.isNotNull(attestationResult.value.sgxCollateral);
-      assert.isDefined(attestationResult.value.runtimeClaims);
-      assert.isUndefined(attestationResult.value.enclaveHeldData);
+      assert.isNotNull(attestationResult.body.sgxCollateral);
+      assert.isDefined(attestationResult.body.runtimeClaims);
+      assert.isUndefined(attestationResult.body.enclaveHeldData);
 
       // Confirm that the JWK response out of the service makes sense and contains
       // the JSON inside the binaryRuntimeData.
-      assert.isDefined(attestationResult.value.runtimeClaims.jwk);
-      assert.isDefined(attestationResult.value.runtimeClaims.jwk.crv);
-      expect(attestationResult.value.runtimeClaims.jwk.crv).is.equal("P-256");
+      assert.isDefined(attestationResult.body.runtimeClaims.jwk);
+      assert.isDefined(attestationResult.body.runtimeClaims.jwk.crv);
+      expect(attestationResult.body.runtimeClaims.jwk.crv).is.equal("P-256");
 
       assert(attestationResult.token, "Expected a token from the service but did not receive one");
     }
@@ -289,10 +289,10 @@ describe("[AAD] Attestation Client", function() {
         }
       );
 
-      assert.isNotNull(attestationResult.value.sgxCollateral);
-      assert.isUndefined(attestationResult.value.runtimeClaims);
-      expect(attestationResult.value.enclaveHeldData?.length).is.equal(binaryRuntimeData.length);
-      expect(attestationResult.value.enclaveHeldData).to.deep.equal(binaryRuntimeData);
+      assert.isNotNull(attestationResult.body.sgxCollateral);
+      assert.isUndefined(attestationResult.body.runtimeClaims);
+      expect(attestationResult.body.enclaveHeldData?.length).is.equal(binaryRuntimeData.length);
+      expect(attestationResult.body.enclaveHeldData).to.deep.equal(binaryRuntimeData);
 
       assert(attestationResult.token, "Expected a token from the service but did not receive one");
     }
@@ -308,15 +308,15 @@ describe("[AAD] Attestation Client", function() {
         }
       );
 
-      assert.isNotNull(attestationResult.value.sgxCollateral);
-      assert.isDefined(attestationResult.value.runtimeClaims);
-      assert.isUndefined(attestationResult.value.enclaveHeldData);
+      assert.isNotNull(attestationResult.body.sgxCollateral);
+      assert.isDefined(attestationResult.body.runtimeClaims);
+      assert.isUndefined(attestationResult.body.enclaveHeldData);
 
       // Confirm that the JWK response out of the service makes sense and contains
       // the JSON inside the binaryRuntimeData.
-      assert.isDefined(attestationResult.value.runtimeClaims.jwk);
-      assert.isDefined(attestationResult.value.runtimeClaims.jwk.crv);
-      expect(attestationResult.value.runtimeClaims.jwk.crv).is.equal("P-256");
+      assert.isDefined(attestationResult.body.runtimeClaims.jwk);
+      assert.isDefined(attestationResult.body.runtimeClaims.jwk.crv);
+      expect(attestationResult.body.runtimeClaims.jwk.crv).is.equal("P-256");
 
       assert(attestationResult.token, "Expected a token from the service but did not receive one");
     }